refactor(plan-lifecycle): migrate all remaining callers to facade
Redirect all external imports of plan-detection.ts, frontmatter.ts,
receipt.ts, global-state.ts, and pipeline-state.ts through the
plan-lifecycle.ts facade. This covers CLI wrappers (plan-detection-cli,
frontmatter-cli, receipt-cli), secondary callers (pr-lifecycle,
completion-gate, prompt, issue-lifecycle, config, config-cmd,
self-update, repos), TUI components/hooks, interactive helpers,
worktree selection, and all associated tests.

No file outside the five underlying modules now imports directly
from them — all access goes through src/plan-lifecycle.ts.
