+++ src/config-cmd.ts
@@ -19,7 +19,7 @@ import {
 } from "./config.ts";
 import { formatShowConfig } from "./show-config.ts";
 import { runCheck, SUPPORTED_CAPABILITIES } from "./check.ts";
-import { getRepoPipelineDirs } from "./global-state.ts";
+import { getRepoPipelineDirs } from "./plan-lifecycle.ts";
 import { RESET, DIM, TEXT } from "./utils.ts";
 
 // ---------------------------------------------------------------------------
