+++ src/dry-run-safety.test.ts
@@ -18,7 +18,7 @@ import { tmpdir } from "os";
 import { execSync } from "child_process";
 
 import { runRunner, type RunnerOptions } from "./runner.ts";
-import { getRepoPipelineDirs } from "./global-state.ts";
+import { getRepoPipelineDirs } from "./plan-lifecycle.ts";
 import { buildIssuePlanContent } from "./issue-lifecycle.ts";
 import { makeTestResolvedConfig } from "./test-utils.ts";
