+++ src/interactive/run-actions.test.ts
@@ -7,8 +7,8 @@
  */
 
 import { describe, it, expect } from "bun:test";
-import type { PipelineState } from "../pipeline-state.ts";
-import type { BacklogPlan } from "../pipeline-state.ts";
+import type { PipelineState } from "../plan-lifecycle.ts";
+import type { BacklogPlan } from "../plan-lifecycle.ts";
 import {
   unmetDependencies,
   findNextPlanName,
