+++ src/run-target.test.ts
@@ -15,7 +15,7 @@ import { join } from "path";
 import { execSync } from "child_process";
 import { runCliInProcess, useTempGitDir } from "./test-utils.ts";
 import { getConfigFilePath } from "./config.ts";
-import { getRepoPipelineDirs } from "./global-state.ts";
+import { getRepoPipelineDirs } from "./plan-lifecycle.ts";
 
 describe("ralphai run <target>", () => {
   const ctx = useTempGitDir();
