+++ src/tui/components/header.test.ts
@@ -11,7 +11,7 @@
  */
 
 import { describe, it, expect } from "bun:test";
-import type { PipelineState } from "../../pipeline-state.ts";
+import type { PipelineState } from "../../plan-lifecycle.ts";
 import {
   buildHeaderParts,
   buildStalledWarning,
