+++ src/tui/components/header.tsx
@@ -12,7 +12,7 @@
 import { useMemo } from "react";
 import { Box, Text } from "ink";
 
-import type { PipelineState } from "../../pipeline-state.ts";
+import type { PipelineState } from "../../plan-lifecycle.ts";
 import { stalledPlans } from "../../interactive/pipeline-actions.ts";
 
 // ---------------------------------------------------------------------------
