+++ src/tui/components/screen-frame.test.ts
@@ -9,7 +9,7 @@
  */
 
 import { describe, it, expect } from "bun:test";
-import type { PipelineState } from "../../pipeline-state.ts";
+import type { PipelineState } from "../../plan-lifecycle.ts";
 import {
   borderColor,
   bannerColor,
