+++ src/tui/screens/reset.tsx
@@ -24,8 +24,8 @@ import type {
 } from "../components/selectable-list.tsx";
 import { SelectableList } from "../components/selectable-list.tsx";
 import type { DispatchResult } from "../types.ts";
-import type { InProgressPlan } from "../../pipeline-state.ts";
-import type { LivenessStatus } from "../../pipeline-state.ts";
+import type { InProgressPlan } from "../../plan-lifecycle.ts";
+import type { LivenessStatus } from "../../plan-lifecycle.ts";
 
 // ---------------------------------------------------------------------------
 // Types
